Reduce time for building vignette
It took a long time to compile a vignette. Now, the compiling time has been made shorter by saving the output of some lines of code in the vignette and loading the output rather than running the code each time the vignette is compiled. This modification will ease updating the package, both on the developer and the user side.